+++ mayan/settings/base.py
@@ -147,6 +147,7 @@ INSTALLED_APPS = (
     'mayan.apps.file_metadata_exif.apps.FileMetadataEXIFApp',
     'mayan.apps.file_metadata_msg.apps.FileMetadataMSGApp',
     'mayan.apps.file_metadata_ollama.apps.FileMetadataOllamaApp',
+    'mayan.apps.file_metadata_openai.apps.FileMetadataOpenAIApp',
     'mayan.apps.file_metadata_pypdf.apps.FileMetadataPyPDFApp',
     'mayan.apps.linking.apps.LinkingApp',
     'mayan.apps.mailer.apps.MailerApp',