+++ HISTORY.rst
@@ -143,6 +143,7 @@
   subclass execution and avoid having that logic in the task.
 - Update `FormFieldMixinFilteredQueryset` to load content from
   `source_queryset' and not `queryset`.
+- Add file metadata driver for OpenAI.
 
 4.9.5 (XXXX-XX-XX)
 ==================